requirements.txt
amqp==5.2.0
asgiref==3.8.1
billiard==4.2.0
build==1.2.2.post1
celery==5.4.0
certifi==2024.8.30
cffi==1.17.1
charset-normalizer==3.3.2
click==8.1.7
click-didyoumean==0.3.1
click-plugins==1.1.1
click-repl==0.3.0
colorama==0.4.6
cryptography==43.0.1
Django==5.1.4
django-appconf==1.0.6
django-cors-headers==4.4.0
django-dramatiq==0.11.6
django-encrypted-model-fields==0.6.5
django-extensions==3.2.3
django-filter==24.3
django-grappelli==4.0.1
django-prometheus==2.3.1
django-redis==5.4.0
djangorestframework==3.15.2
djangorestframework-simplejwt==5.3.1
dramatiq==1.17.0
flower==2.0.1
geographiclib==2.0
humanize==4.10.0
idna==3.9
kombu==5.4.1
packaging==24.2
pika==1.3.2
pillow==10.4.0
pip-tools==7.4.1
prometheus_client==0.21.0
prompt_toolkit==3.0.47
pycparser==2.22
pycryptodome==3.20.0
pydotplus==2.0.2
PyJWT==2.9.0
pyotp==2.9.0
pyparsing==3.1.4
pyproj==3.7.0
pyproject_hooks==1.2.0
python-dateutil==2.9.0.post0
python-decouple==3.8
pytz==2024.2
redis==5.0.8
requests==2.32.3
sentry-sdk==2.19.0
setuptools==75.6.0
six==1.16.0
sqlparse==0.5.1
tornado==6.4.1
tzdata==2024.1
urllib3==2.2.3
vine==5.1.0
wcwidth==0.2.13
wheel==0.45.1
